Firebase Logout fixed
Authentication state is now persisted after page reloading
