+++ src/index.js
@@ -27,7 +27,9 @@ import './js/pagination';
 import "./js/clock-animation";
 
 // Firebase
-import './js/firebaseModals';
-import './js/firebaseInit';
-import './js/firebaseAuth';
-import './js/firebaseRegister';
+import './js/firebase/firebaseModals';
+import './js/firebase/firebaseInit';
+import './js/firebase/firebaseAuth';
+import './js/firebase/firebaseRegister';
+import './js/firebase/firebaseCheckLoginStatus';
+import './js/firebase/firebaseLogout';
